First arrangement on Menu and submenus displaying
A lot of work on how the menus horizontal are sometimes not appearing.

Next, A lot of works to arrange the contents of submenu depending on the Menu. Projects have its styles (clicking on the image of project will lead to a new related website apps, News (as a mini-website where you will have a sticker). About NdaY will define the non-permanent Team driven by the GM, Coders, Physicians, Agronomes and so one.