Submenu Contents synchronized with Dynamic Background
Ensure Horizontal Menu and its child trigger the hide of dynamic background.
Submenu contents redesigned and having the link to the projects website (to be build now)

Next improvement the colors used for text and background should relay on the NdaY' Logo. Ask Canvas on what colors can be used for.